requirements.txt
absl-py==2.4.0
asttokens==3.0.1
astunparse==1.6.3
certifi==2026.2.25
charset-normalizer==3.4.5
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
flatbuffers==25.12.19
fonttools==4.61.1
gast==0.7.0
google-pasta==0.2.0
grpcio==1.78.0
h5py==3.14.0
idna==3.11
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
keras==3.13.2
kiwisolver==1.4.9
libclang==18.1.1
markdown-it-py==4.0.0
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
nest-asyncio==1.6.0
numpy==2.4.2
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==3.0.0
pandas-stubs==3.0.0.260204
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.2
prompt_toolkit==3.0.52
protobuf==7.34.0
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pyzmq==27.1.0
requests==2.32.5
rich==14.3.3
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
setuptools==82.0.1
six==1.17.0
stack-data==0.6.3
tensorflow_cpu==2.21.0
termcolor==3.3.0
threadpoolctl==3.6.0
tornado==6.5.4
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
wheel==0.46.3
wrapt==2.1.2
